+++ src/App.jsx
@@ -3,11 +3,10 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { AuthContextProvider } from "./context/AuthContext.jsx";
 
 //Define your routes for APP here
-import LoginScreen from "./pages/login/LoginScreen";
-import AdminDashoboard from "./pages/adminDashoard/AdminDashboard";
+import LoginPage from "./pages/login/LoginPage.jsx";
 import ChangePassword from "./pages/ChangePasswordPage/ResetPasswordPage";
 import ForgotPassword from "./pages/ForgotPage/ForgotPage";
-
+import Applicantions from "./pages/applications/Applications.jsx";
 //Private Routes will be wrapped in below component
 import PrivateRoute from "./routes/PrivateRoute";
 
@@ -16,12 +15,12 @@ const App = () => {
     <AuthContextProvider>
       <BrowserRouter>
         <Routes>
-          <Route path="/login" element={<LoginScreen />} />
+          <Route path="/login" element={<LoginPage />} />
           <Route path="/forgot/password" element={<ForgotPassword />} />
           <Route path="/change/password" element={<ChangePassword />} />
           {/* Private Route, can't access without token */}
           <Route path="/batch/:id/applications" element={<PrivateRoute />}>
-            <Route path="/batch/:id/applications" element={<AdminDashoboard />} />
+            <Route path="/batch/:id/applications" element={<Applicantions />} />
           </Route>
         </Routes>
       </BrowserRouter>
