--- src/components/Button.jsx
@@ -1,8 +1,0 @@
-//Remove this file
-
-//keep components  like button, Modal, input fields etc in this folder
-import React from "react";
-
-export const Button = ({buttonText}) => {
-  return  <button className="login-btn">{buttonText}</button>;
-};
