+++ src/components/PasswordRequirement.jsx
@@ -20,11 +20,11 @@ export const checkPasswordCriteria = (password,setpasswordCriteria) => {
   let criteria = passwordRequirements(password);
       const content = (
         <ul>
-          <li>{criteria.lengthCheck ? '✅' : '❌'} Minimum 8 characters</li>
-          <li>{criteria.upperCaseCheck ? '✅' : '❌'} At least one uppercase letter</li>
-          <li>{criteria.lowerCaseCheck ? '✅' : '❌'} At least one lowercase letter</li>
-          <li>{criteria.numberCheck ? '✅' : '❌'} At least one digit</li>
-          <li>{criteria.specialCharCheck ? '✅' : '❌'} At least one special character</li>
+          <li className='criteria'>{criteria.lengthCheck ? <span class="material-symbols-outlined tick">check_circle</span> : <span class="material-symbols-outlined cancel">cancel</span>} Minimum 8 characters</li>
+          <li className='criteria'>{criteria.upperCaseCheck ? <span class="material-symbols-outlined tick">check_circle</span> : <span class="material-symbols-outlined cancel">cancel</span>} At least one uppercase letter</li>
+          <li className='criteria'>{criteria.lowerCaseCheck ? <span class="material-symbols-outlined tick">check_circle</span> : <span class="material-symbols-outlined cancel">cancel</span>} At least one lowercase letter</li>
+          <li className='criteria'>{criteria.numberCheck ? <span class="material-symbols-outlined tick">check_circle</span> : <span class="material-symbols-outlined cancel">cancel</span>} At least one digit</li>
+          <li className='criteria'>{criteria.specialCharCheck ? <span class="material-symbols-outlined tick">check_circle</span> : <span class="material-symbols-outlined cancel">cancel</span>} At least one special character</li>
         </ul>
       );
       setpasswordCriteria(content);
