+++ src/components/TableView.jsx
@@ -1,0 +1,30 @@
+import React, { useState } from 'react';
+const TableComponent = () => {
+
+ 
+  return (
+    <table className="antd-table">
+      <thead>
+        <tr>
+          <th>Id</th>
+          <th>First Name</th>
+          <th>Last Name</th>
+          <th>Address</th>
+          <th>DOB</th>
+          <th>Email</th>
+          <th>invite</th>
+          <th>view more</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr>
+          <td>tamil</td>
+        </tr>
+      </tbody>
+    </table>
+  );
+};
+
+
+
+export default TableComponent;
