+++ src/layouts/Navbar.jsx
@@ -1,8 +1,7 @@
-import React, { useState } from 'react';
-import { Popover, Button } from 'antd';
-
-
-const Profile = () => {
+// Navbar.js
+import React from 'react';
+import { Button,Popover } from 'antd';
+const Navbar = () => {
   const content = (
     <div>
       <p>
@@ -15,15 +14,21 @@ const Profile = () => {
       </Button>
     </div>
   );
+
   return (
-    <div className='avatar__container'>
+    <div className="navbar">
+      <div>
+        <h2>Welcome to Vignesh Selvaraj</h2>
+      </div>
+      <div className='avatar__container'>
        <Popover content={content} placement="bottomRight" trigger="click">
           <div className='avatar__logo'>
            <span className="material-symbols-outlined">account_circle</span>
           </div>
         </Popover>
     </div>
+    </div>
   );
 };
 
-export default Profile;
+export default Navbar;
