--- src/layouts/SampleNavBar.jsx
@@ -1,16 +1,0 @@
-//Remove this file
-
-import { SampleButton } from "../components/SampleButton";
-
-
-// you may use components inside layout components
-
-export const SampleNavBar = () => {
-  return (
-    <div>
-      Sample nav bar
-      <SampleButton />
-    </div>
-  );
-};
-
