--- src/pages/adminDashoard/components/Navbar.jsx
@@ -1,15 +1,0 @@
-// Navbar.js
-import React from 'react';
-import Profile from './Profile';
-const Navbar = () => {
-  return (
-    <div className="navbar">
-      <div>
-        <h2>Welcome to Vignesh Selvaraj</h2>
-      </div>
-      <Profile/>
-    </div>
-  );
-};
-
-export default Navbar;
