+++ src/pages/applications/Applicantions.css
@@ -1,4 +1,4 @@
-/* App.css */
+
 * {
   padding: 0;
   margin: 0;
@@ -99,23 +99,7 @@ body {
   padding: 20px;
 }
 
-/* .filter{
-  margin: 10px 5px;
-  padding: 10px;
-  display: none;
-}
-.filter p{
-  padding-bottom: 5px;
-}
-.filter.show{
-  display: block;
-} */
-
-/* a{
-  text-decoration: none;
-  color: white;
-} */
-.search__bar{
+.header-controls .search__bar{
   max-width: 200px;
   width: 100%;
 }
