+++ src/pages/login/LoginPage.jsx
@@ -1,20 +1,16 @@
 import React, { useState } from "react";
-import "./LoginScreen.css";
+import "./LoginPage.css";
 import { Popover } from "antd";
 import { useNavigate } from "react-router-dom";
 import { validate } from "../../utils/validate";
-import { Button } from "../../components/Button";
 import {
   checkPasswordCriteria,
   passwordRequirements,
 } from "../../components/PasswordRequirement";
-import UserService from "../../services/UserService";
-import DckapPalliLogo from "/images/dckap_palli_logo_lg.png";
-import ManagerLoginLogo from "/images/manager_login_logo.png";
+import dckapPalliLogon from "../../../public/images/dckap_palli_logo_lg.svg"
+import managerLoginLogo from "../../../public/images/manager_login_image.svg"
 import Input from "../../components/Input";
 import { API_END_POINT } from "../../../config";
-
-
 import { useAuth } from "../../context/AuthContext";
 import axios from "axios";
 
@@ -98,12 +94,12 @@ const LoginScreen = () => {
           <p>Make Sure Your Account is Secure</p>
         </div>
         <div className="right__side--header">
-          <img src={DckapPalliLogo} alt="" />
+          <img src={dckapPalliLogon} alt="dckap-logo" />
         </div>
       </div>
       <div className="input__containers">
         <div className="left__side__image__container">
-          <img src={ManagerLoginLogo} alt="" />
+          <img src={managerLoginLogo} alt="manager-logo" />
         </div>
         <div className="right__side-input__container">
           <form onSubmit={handleLoginSubmit} className="login-form">
@@ -141,7 +137,7 @@ const LoginScreen = () => {
 
             <a href="/forgot/password">Forgot password ?</a>
             <div className="login__btn__container">
-              <Button buttonText={"login"} />
+            <button className="login-btn">Login</button>
             </div>
           </form>
         </div>
